Update README.md
This badge wasn't sitting well with me when i saw it. I feel like it is a little bit grifty, For sure open to feedback on this but i think it's cleaner if we just remove it. I think being open to contributions is implied in our project readme's where we link to our contributor guidelines